ai-threatintel/requirements.txt
# FastAPI Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
pydantic-settings==2.6.0

# HTTP Clients
httpx==0.27.0
requests==2.32.3
aiohttp==3.11.0

# Data Processing
python-multipart==0.0.17
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Threat Intelligence APIs
shodan==1.31.0
python-whois==0.8.0
dnspython==2.7.0

# AI/ML (if using OpenAI)
openai==1.58.1

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.2

# Logging & Monitoring
structlog==24.4.0

# Date/Time
python-dateutil==2.9.0

# Security
cryptography==44.0.0

# ML Dependencies (for ml_model_service.py)
numpy==2.2.1
scikit-learn==1.6.0
joblib==1.4.2